Dynamic SQL to generate select query on a table along with order by clause on the selected columns
I came up with a problem where I need to compare data in a table before and after execution of a business logic. I had to compare multiple tables, so I started thinking of a way to to generate the select statement along with order by clause. I utilized Microsoft SQL Server’s INFORMATION_SCHEMA.COLUMNS to solve the problem.